+++ src/components/AgreementFilters/AgreementFilters.js
@@ -4,7 +4,7 @@ import { FormattedMessage } from 'react-intl';
 
 import { Accordion, AccordionSet, FilterAccordionHeader, Selection } from '@folio/stripes/components';
 import { CheckboxFilter, MultiSelectionFilter } from '@folio/stripes/smart-components';
-import { OrganizationSelection } from '@folio/stripes-erm-components';
+import { InternalContactSelection, OrganizationSelection } from '@folio/stripes-erm-components';
 
 const FILTERS = [
   'agreementStatus',
@@ -136,10 +136,34 @@ export default class AgreementFilters extends React.Component {
     );
   }
 
+  renderInternalContactFilter = () => {
+    const { activeFilters } = this.props;
+    const contactFilters = activeFilters.contacts || [];
+
+    return (
+      <Accordion
+        closedByDefault
+        displayClearButton={contactFilters.length > 0}
+        header={FilterAccordionHeader}
+        id="internal-contacts-filter"
+        label={<FormattedMessage id="ui-agreements.agreements.internalContacts" />}
+        onClearFilter={() => this.props.filterHandlers.clearGroup('contacts')}
+        separator={false}
+      >
+        <InternalContactSelection
+          path="erm/contacts"
+          id="agreement-internal-contacts-filter"
+          input={{
+            name: 'agreement-contacts-filter',
+            onChange: value => this.props.filterHandlers.state({ ...activeFilters, contacts: [value] }),
+            value: contactFilters[0] || '',
+          }}
+        />
+      </Accordion>
+    );
+  }
+
   renderTagsFilter = () => {
-    // const tags = get(this.props.data, 'tagValues.records', []);
-    // TODO: TEST USING THE VALUES GENERATED IN GDSFP
-    // const dataOptions = tags.map(({ label }) => ({ value: label, label }));
     const { activeFilters } = this.props;
     const tagFilters = activeFilters.tags || [];
 
@@ -172,6 +196,7 @@ export default class AgreementFilters extends React.Component {
         {this.renderCheckboxFilter('isPerpetual', { closedByDefault: true })}
         {this.renderOrganizationFilter()}
         {this.renderOrganizationRoleFilter()}
+        {this.renderInternalContactFilter()}
         {this.renderTagsFilter()}
       </AccordionSet>
     );